==> 0_LI_PM_data_cleaning.ipynb <==
import numpy as np
import pandas as pd
import matplotlib
import matplotlib.pyplot as plt
%matplotlib inline

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)

### LOADING, CLEANING AND FILTERING DATA

df1 = pd.read_csv('2023-01-23_job_details.csv', delimiter = ',')
df2 = pd.read_csv('2023-02-11_job_details.csv', delimiter = ',')
df3 = pd.read_csv('2023-05-06_job_details.csv', delimiter = ',')
df4 = pd.read_csv('2023-05-30_job_details.csv', delimiter = ',')
df5 = pd.read_csv('2023-07-05_job_details.csv', delimiter = ',')

df = pd.concat([df1, df2, df3, df4, df5])

df.shape

df = df.drop_duplicates()
df.shape

df

df.describe(include = 'all')

#find and drop rows where jobDescription is NaN

df[pd.isna(df['jobDescription'])] 
#1, 157, 160, 202, 214

df.drop(df.index[[1, 157, 160, 202, 214]], inplace=True)

df.shape

#find and drop rows where jobDescriptions are duplicated

duplicated_rows = df.duplicated(subset='jobDescription')

df = df.drop_duplicates(subset='jobDescription')
df.describe(include = 'all')

#drop unnecessary columns

df.drop(columns=['companyUrl', 'viewsCount', 'companyDescription', 'companyLogoUrl', 'jobFunctions',
                'jobState', 'timestamp', 'query', 'error', 'jobPosterProfileUrl', 'jobPosterName'],
        inplace=True)

df.shape

df.dtypes

#converting data types of columns 

nan_indices = df['companyStaffCount'].isna()
df.loc[nan_indices, 'companyStaffCount'] = 0

df = df.astype({'jobId': 'int',
                'companyStaffCount': 'int',
                'remoteAllowed': 'bool'})

df['postedAt'] = pd.to_datetime(df['postedAt'])
df['appliesClosedAt'] = pd.to_datetime(df['appliesClosedAt'])


df

#filtering data based on jobTitle

relevant_expressions = ['Product Manager', 'product manager', 'Product Owner', 'Head of .* Product', 'Director of .* Product', 
                        'Product Management', 'Product Analyst', 'Team Lead', 'Product Development']
pattern = '|'.join(relevant_expressions)

relevant_jobAds = df[df['jobTitle'].str.contains(pattern, case=False, na=False, regex=True)]

relevant_jobAds.reset_index(drop=True, inplace=True)
relevant_jobAds

relevant_jobAds.shape

relevant_jobAds.to_csv('relevant_jobAds.csv', index=False)
