==> 1_LI_PM_jobad_basic_analyses.ipynb <==
import numpy as np
import pandas as pd
import matplotlib
import matplotlib.pyplot as plt
%matplotlib inline

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)

## Load data

LI_PM_jobad = pd.read_csv('relevant_jobAds.csv', delimiter = ',')
LI_PM_jobad

## Informations about companies

#location of work

LI_PM_jobad.groupby('jobLocation').count().jobId.sort_values(ascending = False)

#data cleaning in 'jobLocation' column

LI_PM_jobad['jobLocation'] = LI_PM_jobad['jobLocation'].replace(['Budapest Metropolitan Area', 'Budapest, Budapest, Hungary'], 'Budapest, Hungary')
LI_PM_jobad['jobLocation'] = LI_PM_jobad['jobLocation'].replace('EMEA', 'Europe-MiddleEast-Africa Region')
LI_PM_jobad['jobLocation'] = LI_PM_jobad['jobLocation'].replace('Berlin, Germany', 'Germany')


#company staff count

tmp = LI_PM_jobad.sort_values(by='companyStaffCount', ascending=False)
tmp = tmp[['companyName', 'companyStaffCount']].reset_index(drop=True)
tmp = tmp.drop_duplicates().reset_index(drop=True)
tmp

#companyName correction in row_139
LI_PM_jobad.loc[139, 'companyName'] = 'Jófogás'

#job industries

pd.set_option('display.max_colwidth', None)

tmp = LI_PM_jobad[['companyName', 'jobIndustries']]

tmp['jobIndustries'] = tmp['jobIndustries'].replace(np.nan, 'Missing values')

tmp = tmp.sort_values(by = 'jobIndustries').drop_duplicates().reset_index(drop=True)

tmp

## Type of workplace and employment

#type of workplace

LI_PM_jobad['workplaceType'] = LI_PM_jobad['workplaceType'].replace(np.nan, 'Missing values')

LI_PM_jobad.groupby('workplaceType').count().jobId.sort_values(ascending = False)

#type of employment

LI_PM_jobad['jobType'] = LI_PM_jobad['jobType'].replace(np.nan, 'Missing values')

LI_PM_jobad.groupby('jobType').count().jobId.sort_values(ascending = False)

## Seniority level based on LI classification

tmp = LI_PM_jobad[['companyName', 'experienceLevel']]

tmp['experienceLevel'] = tmp['experienceLevel'].replace(np.nan, 'Missing values')

tmp

tmp.groupby('experienceLevel').count().companyName.sort_values(ascending = False)

## Applicants count

tmp = LI_PM_jobad[['companyName', 'applicantsCount']]

tmp['applicantsCount'] = tmp['applicantsCount'].replace(np.nan, 0)

tmp.sort_values(by = 'applicantsCount', ascending = False).drop_duplicates().reset_index(drop=True)

## How long is an advertisement active?

tmp = LI_PM_jobad[['companyName','appliesClosedAt','postedAt']]

tmp['postedAt'] = pd.to_datetime(tmp['postedAt']).dt.round('D')
tmp['appliesClosedAt'] = pd.to_datetime(tmp['appliesClosedAt']).dt.round('D')

tmp['jobad_active_was'] = (tmp['appliesClosedAt'] - tmp['postedAt']).dt.days

tmp = tmp.sort_values(by = ['jobad_active_was'], ascending = False)

tmp = tmp[['companyName', 'jobad_active_was']].drop_duplicates().reset_index(drop = True)
tmp

## Creating a filtered csv for further analysis

LI_PM_jobad[['companyName', 'jobDescription']].to_csv('LI_PM_jobDesc.csv', index=False)
