==> 6_LI_PM_benefitsPerks_analysis.ipynb <==
import numpy as np
import pandas as pd
import nltk
import time
import openai
openai.api_key = "TYPE HERE YOUR API KEY"

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)

## Benefits and perks

#Create a new dataframe
LI_PM_jobad = pd.read_csv('LI_PM_jobDesc.csv', delimiter = ',')
df = pd.DataFrame()
df['companyName'] = LI_PM_jobad['companyName']
df['jobDescription'] = LI_PM_jobad['jobDescription']
df['benefitsPerks'] = ''

#iterate through the dataframe
for index, row in df.iterrows():
    
    prompt_benefitsPerks = row['jobDescription'] + "/n --- /n This is a product manager job description.\
         Make a list of benefits and additional perks offered by employers, that are included in the description!\
         Focus on benefits and perks related to finances and working conditions!\
         Don't use whole sentences, just list keywords in your answers!\
         Answer the question as truthfully as possible using the provided text.\
         If there is no clear information about benefits and perks in the description,\
         use only this value: 'Unknown' and don't use whole sentences."

    response_benefitsPerks = openai.ChatCompletion.create(
        model = "gpt-4",
        messages = [
            {"role": "system", "content": "You are a helpful data science assistant."},
            {"role": "user", "content": prompt_benefitsPerks}],
        max_tokens=100,
        temperature=0)
    
    message_benefitsPerks = response_benefitsPerks['choices'][0]['message']['content'].strip()

    df.loc[index, 'benefitsPerks'] = message_benefitsPerks
    
    time.sleep(0.5)
        
    if index % 10 == 0:
        print(f"Processed row {index} for company {row['companyName']} with benefits and perks: {message_benefitsPerks}")
        
#print results
df[['companyName', 'benefitsPerks']]

## Further analysis with chatGPT

At first, I classified the text fragments collected by the AI from the 212 job advertisements into 5 large groups. Then, I copied these groups one by one to ChatGPT and asked it to help me with the analysis using the following prompt:

"I will show some job advertisement details about benefits and perks given by employers - from now on, I will refer to this group of advertisements as jobAd_n.
Tasks:
1) Combine similar benefits and perks in the jobAd_n advertisement group!
2) Then, collect the 15 most frequently occurring benefits and perks from this jobAd_n group!
3) Finally, make a table of how many advertisements appeared in the jobAd_n group!"

Then, I saved the table created by ChatGPT for myself and performed the same analysis with all advertisement groups. Finally, I copied the tables created after all 5 groups and asked ChatGPT to combine similar benefits and perks and make a last combined table.
